Added Heap Project Files
The heap project I worked on in CSE 373
Integrate multiple ADTs and data structures to solve a problem with complex requirements.

Due to the specific runtime requirements of this heap implementation, our solution will need to creatively combine different ADTs and data structures to design a better solution.

Plan the implementation of a data structure and its invariants.

At a high level, it’s necessary to determine the invariants that a data structure should maintain before and after every method call, and at a low level, it’s useful to determine how these invariants affect the shared functionality that can be extracted into helper methods. In the process, it may also be necessary to consider which of the invariants these helper methods themselves require in order to function correctly.